Update index.html
Nevermind, the double period's don't work for the index, and only work for the other pages. Great to know and I am glad I figured it out. It was just a guess because I noticed that the index was my only html page that linked to the css with a single period. So when I added a double, it messed up the style. And when I removed it from all the links, they started to work. So, I hope it works here too.